pyproject.toml
[project]
name = "collyzer"
version = "0.1.0"
description = "Collect and analyze logs"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "paramiko>=3.5.1",
    "python-decouple>=3.8",
    "pyyaml>=6.0.2",
    "redis>=6.2.0",
    "sqlalchemy>=2.0.41",
]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "ruff>=0.12.3",
]

[tool.pytest.ini_options]
pythonpath = [".", "src"]
